+++ internal/core/cluster_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/sapcc/limes"
 )
 
-//nolint:dupl
 func TestClusterReportRender(t *testing.T) {
 	mockJSONBytes, err := fixtureBytes("cluster-get-west.json")
 	th.AssertNoErr(t, err)
@@ -42,43 +41,4 @@ func TestClusterReportRender(t *testing.T) {
 	err = RenderReports(opts, rep).Write(&actual)
 	th.AssertNoErr(t, err)
 	assertEquals(t, "cluster-get-west.csv", actual.Bytes())
-}
-
-func TestClusterReportsRender(t *testing.T) {
-	type listData struct {
-		CurrentCluster string                `json:"current_cluster"`
-		Clusters       []limes.ClusterReport `json:"clusters"`
-	}
-
-	// List
-	mockJSONBytes, err := fixtureBytes("cluster-list.json")
-	th.AssertNoErr(t, err)
-	var data listData
-	err = json.Unmarshal(mockJSONBytes, &data)
-	th.AssertNoErr(t, err)
-
-	opts := &OutputOpts{
-		CSVRecFmt: CSVRecordFormatDefault,
-		Humanize:  false,
-	}
-	var actual bytes.Buffer
-	reps := LimesClustersToReportRenderer(data.Clusters)
-	err = RenderReports(opts, reps...).Write(&actual)
-	th.AssertNoErr(t, err)
-	assertEquals(t, "cluster-list.csv", actual.Bytes())
-
-	// Filtered list with long CSV format and human-readable values
-	mockJSONBytes, err = fixtureBytes("cluster-list-filtered.json")
-	th.AssertNoErr(t, err)
-	var filteredData listData
-	err = json.Unmarshal(mockJSONBytes, &filteredData)
-	th.AssertNoErr(t, err)
-
-	opts.CSVRecFmt = CSVRecordFormatLong
-	opts.Humanize = true
-	actual.Reset()
-	reps = LimesClustersToReportRenderer(filteredData.Clusters)
-	err = RenderReports(opts, reps...).Write(&actual)
-	th.AssertNoErr(t, err)
-	assertEquals(t, "cluster-list-filtered.csv", actual.Bytes())
 }