+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated all dependencies to their latest version.
 
+### Removed
+
+- `cluster list` command; Limes has removed multi-cluster support.
+
 ## 2.0.1 - 2021-10-06
 
 ### Fixed